Renamed to CacheController and added documentation.
